+++ Tests/DependencyInjection/Factory/AbstractStorageFactoryTest.php
@@ -90,10 +90,7 @@ class AbstractStorageFactoryTest extends \PHPUnit\Framework\TestCase
         ));
     }
 
-    /**
-     * @return AbstractStorageFactory|MockObject
-     */
-    protected function createAbstractStorageFactory(): AbstractStorageFactory|MockObject
+    protected function createAbstractStorageFactory(): AbstractStorageFactory&MockObject
     {
         return $this->getMockForAbstractClass(AbstractStorageFactory::class);
     }