+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## 3.0.0 (2022-09-xx)
 
-* Support for Symfony <5.4 dropped
-* Support for PHP < 8 dropped
+* Support for Symfony < 6.1 dropped
+* Support for PHP < 8.1 dropped
 * remove deprecated functions getPayum, getHttpRequestVerifier and getTokenFactory in PayumController
 
 ## 2.5.0 (2022-07-xx)
 
-* Support for Symfony 4.0 - 4.3 dropped
+* Support for Symfony 5.0 - 5.3 dropped
 * Added support for Symfony 6.0
 * Minimum Payum Core dependency updated to 1.7.2